+++ LQCD_Software_Dependent_Code/openQCD-FASTSUM/SimulationsStatusInformation.bash
@@ -47,6 +47,7 @@ function CreateOutputFileInTheStandardFormat_openQCD-FASTSUM()
     fi
     # NOTE: "N.A." is printed for information that openQCD does not give
     #       "nan" are printed for not found numbers -> https://stackoverflow.com/a/23622339
+    #       Plaquette and Polyakov loop values are divided by Nc=3 since not done in the code!
     awk -v offset="${trShift}"\
         '
         function PrintLineToFile(tr, plaq, L_re, L_im, L_norm, dH, acc, time)
@@ -58,7 +59,7 @@ function CreateOutputFileInTheStandardFormat_openQCD-FASTSUM()
             if(dH == ""){dH="nan"}
             if(acc == ""){acc="nan"}
             if(time == ""){time="nan"}
-            printf "%8d%25s%10s%10s%25s%25s%25.15f%15s%6d%10.1f\n", tr, plaq, "N.A.", "N.A.", L_re, L_im, L_norm, dH, acc, time
+            printf "%8d%25.15f%10s%10s%25.15f%25.15f%25.15f%15s%6d%10.1f\n", tr, plaq, "N.A.", "N.A.", L_re, L_im, L_norm, dH, acc, time
         }
         {
             if( $0 ~ /^Trajectory no [1-9][0-9]*$/ )
@@ -79,14 +80,14 @@ function CreateOutputFileInTheStandardFormat_openQCD-FASTSUM()
             }
             if( $0 ~ /^Average plaquette =/ )
             {
-                plaquette=$4
+                plaquette=$4/3.0
                 next
             }
             if( $0 ~ /^Polyakov loop =/ )
             {
-                polyRe=$4
-                polyIm=$5
-                polyNorm=sqrt($4**2+$5**2)
+                polyRe=$4/3.0
+                polyIm=$5/3.0
+                polyNorm=sqrt(polyRe**2+polyIm**2)
                 next
             }
             if( $0 ~ /^Time per trajectory =/ )